refactor(webapp): unify Button + route L0 primitives through leaf cn (A1)
Unify the button primitive into the canonical L0 file (preserve-both), and fix
an app-graph coupling the shims introduced.

- components/ui/button.tsx now holds the app button verbatim as the bare
  `Button`/`buttonVariants`/`ButtonProps` (119 app call-sites untouched) plus the
  widget button under `ButtonWidget`/`buttonWidgetVariants`/`ButtonWidgetProps`
  (+ the `ButtonVariant` union). @/widgets/components/ui/button is now a shim
  aliasing those back. Class strings verbatim (string-diff vs HEAD clean).
- Fix: the 6 shimmed L0 primitives (avatar, switch, tooltip, skeleton, card,
  button) now import `cn` from the leaf @/lib/cn instead of @/lib/utils. Routing
  widget primitives through @/lib/utils dragged the app/wagmi graph
  (config.default -> createConfig) into the widget layer, which broke widget
  tests that mock wagmi without createConfig (PsmConversionInputs in isolation).
  The leaf cn has zero app deps, honoring the L0 "zero app knowledge" rule.

Verified: tsc --noEmit green; PsmConversionInputs passes in isolation (was
failing); combined 10-file widget/module batch green (79 tests).
