+++ apps/webapp/src/components/ui/button.tsx
@@ -2,7 +2,12 @@ import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
+
+/* =============================================================================
+ * App button (canonical `Button`) — the original components/ui look. UNCHANGED.
+ * Consumed by `@/components/ui/button` call-sites. Pixel-identical to pre-A1.
+ * ========================================================================== */
 
 const buttonVariants = cva(
   'inline-flex items-center justify-center whitespace-nowrap rounded-xl text-sm font-medium ring-offset-background transition-[background-color,background-image,opacity,border-color,color,box-shadow] duration-250 ease-out-expo focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:bg-primaryDisabled disabled:text-surfaceAlt light:disabled:text-textDimmed',
@@ -65,4 +70,81 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
 );
 Button.displayName = 'Button';
 
+/* =============================================================================
+ * Widget button (`ButtonWidget`) — the original widgets/components/ui look,
+ * relocated here under suffixed names per ticket A1's preserve-both decision.
+ * Exposed to widget call-sites via the `@/widgets/components/ui/button` shim,
+ * which aliases `ButtonWidget` -> `Button`, `buttonWidgetVariants` ->
+ * `buttonVariants`, `ButtonWidgetProps` -> `ButtonProps`. The two buttons
+ * diverge on shared variant names (secondary/pill/pagination/link/ghost/input)
+ * and the base radius/disabled rules; convergence is deferred to a follow-up
+ * design-led ticket. Class strings are kept VERBATIM.
+ * ========================================================================== */
+
+export type ButtonVariant =
+  | 'default'
+  | 'secondary'
+  | 'pill'
+  | 'chip'
+  | 'link'
+  | 'pagination'
+  | 'paginationActive'
+  | 'input'
+  | 'primary'
+  | 'primaryAlt'
+  | 'ghost';
+
+const buttonWidgetVariants = cva(
+  'inline-flex items-center justify-center whitespace-nowrap rounded-[12px] text-sm font-medium ring-offset-background transition-[background-color,background-image,opacity,border-color,color,box-shadow] duration-250 ease-out-expo focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none',
+  {
+    variants: {
+      variant: {
+        default:
+          'bg-primary text-text hover:bg-primaryHover active:bg-primaryActive focus:bg-primaryFocus disabled:bg-primaryDisabled',
+        primary:
+          'bg-radial-(--gradient-position) from-primary-start/100 to-primary-end/100 text-text hover:from-primary-start/100 hover:to-primary-end/100 focus:from-primary-start/100 focus:to-primary-end/100 bg-blend-overlay hover:bg-white/10 focus:border-transparent focus:bg-white/15',
+        primaryAlt:
+          'bg-radial-(--gradient-position) from-primary-alt-start/100 to-primary-alt-end/100 border text-text hover:from-primary-alt-start/60 hover:to-primary-alt-end/60 active:from-primary-alt-start/45 active:to-primary-alt-end/45 focus:from-primary-alt-start/45 focus:to-primary-alt-end/45 disabled:from-primary-alt-start/35 disabled:to-primary-alt-end/35',
+        secondary:
+          'bg-transparent text-text border border-textSecondary hover:bg-[rgb(77,76,111)] active:bg-[rgb(94,92,136)] disabled:border-textMuted disabled:text-textMuted',
+        pill: 'bg-primary text-text rounded-full hover:bg-primaryHover active:bg-primaryActive focus:bg-primaryFocus',
+        chip: 'bg-secondary text-text rounded-full hover:bg-secondaryHover active:bg-secondaryActive, focus:bg-secondaryFocus',
+        link: 'text-textSecondary no-underline disabled:text-textMuted',
+        purpleLink: 'text-textEmphasis',
+        pagination:
+          'text-selectActive light:text-textSecondary text-base leading-normal bg-radial-(--gradient-position) from-primary-alt-start/0 to-primary-alt-end/0 rounded-full hover:from-primary-alt-start/50 hover:to-primary-alt-end/50 hover:text-text focus:border-2 focus:border-primaryActive focus:text-text active:text-text active:from-primary-alt-start/30 active:to-primary-alt-end/30 disabled:bg-radial-(--gradient-position) disabled:from-primary-alt-start/0 active:to-primary-alt-end/0 rounded-full! border-0!',
+        paginationActive:
+          'bg-radial-(--gradient-position) from-primary-start/100 to-primary-end/100 hover:bg-primaryHover rounded-full! text-text',
+        input:
+          'bg-black/20 light:bg-surfaceAlt hover:bg-white/10 active:bg-white/7 text-text text-[13px] font-normal leading-4 disabled:pointer-events-auto disabled:cursor-not-allowed font-graphik',
+        ghost: 'text-text hover:text-white/80 active:text-white/60 light:hover:text-text light:active:text-text'
+      },
+      size: {
+        default: 'h-10 px-4 py-2',
+        sm: 'h-9 rounded-full px-2',
+        icon: 'h-10 w-10',
+        input: 'h-6 pt-[5px] pb-[3px] px-2 rounded-[32px]'
+      }
+    },
+    defaultVariants: {
+      variant: 'default',
+      size: 'default'
+    }
+  }
+);
+
+export interface ButtonWidgetProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof buttonWidgetVariants> {
+  asChild?: boolean;
+}
+
+const ButtonWidget = React.forwardRef<HTMLButtonElement, ButtonWidgetProps>(
+  ({ className, variant, size, asChild = false, ...props }, ref) => {
+    const Comp = asChild ? Slot : 'button';
+    return <Comp className={cn(buttonWidgetVariants({ variant, size, className }))} ref={ref} {...props} />;
+  }
+);
+ButtonWidget.displayName = 'Button';
+
 export { Button, buttonVariants };
+export { ButtonWidget, buttonWidgetVariants };
