+++ apps/webapp/src/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 import { cva, type VariantProps } from 'class-variance-authority';
 import { motion } from 'motion/react';
