+++ apps/webapp/src/components/ui/tooltip.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 import { useIsTouchDevice } from '@/hooks';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 
 const TooltipProvider = TooltipPrimitive.Provider;
